+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ### Changed
 
-* **Breaking**: `gouda.display.print_image`, `gouda.display.show_video`, and `gouda.display.print_grid` now use `file_name` instead of `toFile` to specify the optional output file name. The `toFile` argument has been removed.
+* `gouda.display.print_image`, `gouda.display.show_video`, and `gouda.display.print_grid` now use `file_name` instead of `toFile` to specify the optional output file name. The `toFile` argument has been removed.
+* `gouda.file_methods.fast_glob` and `gouda.file_methods.find_images` now use `as_iterator` instead of `iter`.
 
 ### Deprecated